Allow mpv-1.dll to be placed in python script directory
The windows dll search strategy is pathetic. For me at least, it only looks in system32 and the executable directory - not the current directory. This works fine for compiled programs, but for python, the executable is likely to be far away from the script which is being executed. And I don't know about you, but I don't want my python scripts installing their dependencies alongside my system-wide python interpreter. This change allows python scripts using mpv.py to simply ship mpv-1.dll along with them, and mpv.py will attempt to load it for the current working directory.