Actually search sys.path
Here's the root of the issue. `CDLL` doesn't search sys.path. I don't know what it does - the actual logic is native code and I CBA digging into the cpython source to find out. But in practise it only searches system32 and beside python.exe. `find_library` is the tool which we need to use to make this work. I've also added a helpful error message. This change brings the code in line with the advice you added in the README, regarding adding the dll to PATH.

I added abspath to work around a bug I had - when there is a `;;` in PATH, that's interpreted as an empty string. The filename of the library is added on to this, and the result is that `find_library` looks in the current directory. This works for me because that's where I have mpv-1.dll stored, so `find_library` returns the string `mpv-1.dll` - literally what you had originally. I understand this will bring back your concerns about security - however I don't think a normal windows install will have a `;;` in PATH, so most people should not encounter this.